requirements.txt
anyio==4.12.1
cachetools==5.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==41.0.7
google-auth==2.23.4
google-auth-oauthlib==1.1.0
gspread==5.12.4
gspread-formatting==1.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
oauthlib==3.3.1
packaging==26.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
python-dotenv==1.2.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
typing_extensions==4.15.0
urllib3==2.6.3
